assets/ingest_data/requirements.txt
boto3==1.28.64
botocore==1.31.64
langchain==0.1.14
llmsherpa==0.1.4
opensearch-py==2.5.0
pandas==2.1.4
python-dateutil==2.8.2
pytz== 2023.3.post1
pypdf==4.1.0
urllib3==1.26.18